requirements.txt
kafka-python==2.0.2
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
